Moved unknown-prop warning into a devtool
