+++ src/renderers/dom/ReactDOM.js
@@ -15,6 +15,7 @@
 
 var ReactDOMComponentTree = require('ReactDOMComponentTree');
 var ReactDOMDevtools = require('ReactDOMDevtools');
+var ReactDOMUnknownPropertyDevtool = require('ReactDOMUnknownPropertyDevtool');
 var ReactDefaultInjection = require('ReactDefaultInjection');
 var ReactMount = require('ReactMount');
 var ReactPerf = require('ReactPerf');
@@ -136,6 +137,8 @@ if (__DEV__) {
         break;
       }
     }
+
+    ReactDOMDevtools.addDevtool(new ReactDOMUnknownPropertyDevtool());
   }
 }
